// frontend/src/api/model/activityEventDto.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * TeamPrompts.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ActivityEventType } from './activityEventType';
import type { ActivityTargetType } from './activityTargetType';
import type { UserRef } from './userRef';

export interface ActivityEventDto {
  id: string;
  type: ActivityEventType;
  actor: unknown | null | UserRef;
  createdAt: string;
  /** @nullable */
  summary: string | null;
  targetType: unknown | null | ActivityTargetType;
  /** @nullable */
  targetId: string | null;
  /** @nullable */
  targetUserId: string | null;
  /** @nullable */
  model: string | null;
  /** @nullable */
  promptTokens: number | null;
  /** @nullable */
  completionTokens: number | null;
  /** @nullable */
  totalTokens: number | null;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
     */
  costUsd: number | null;
  metadata: string;
}